Checking author when editing and deleting ads

+++ src/app/ads/delete.controller.js
@@ -3,15 +3,20 @@
 
   angular
     .module('angularRails')
-    .controller('AdsDeleteController', function($rootScope, $scope, $state, $stateParams, $auth, $location, Ad) {
+    .controller('AdsDeleteController', function($rootScope, $scope, $state, $stateParams, $auth, $location, Ad, CurrentUser) {
+
 
       $scope.delete_ad = function (adId) {
         if (confirm("Are you sure you want to delete this ad?")){
           Ad.get(adId).then(function(ad){
-            ad.delete().then(function(){
-              $scope.ads = Ad.query();
-              $location.path('/ads');
-            })
+            CurrentUser.query().then(function(user){
+              if (user.id === ad.user.id){
+                ad.delete().then(function(){
+                  $scope.ads = Ad.query();
+                  $location.path('/ads');
+                })
+              }
+            });
           })
         }
       };
